fix(test): scope autouse DB override to e2e and unblock CI
The autouse override fixture in tests/conftest.py forced every test
(unit + integration) to import the FastAPI app and validate Settings,
which broke CI for tests that previously didn't touch config.

- Move _override_app_database to tests/e2e/conftest.py so only e2e
  triggers app import; unit/integration suites stay env-independent.
- CI: add ADMIN_*, DATABASE_ENGINE env vars (now required by Settings)
  and switch ENV from "test" (invalid) to "local"; drop --ignore=tests/e2e
  so the e2e suite is exercised under the new SQLite override path.
